# docs/api/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "BetaV1Result": {
        "description": "The regression half of `RiskV1Result`, for consumers that only need\nvolatility/beta and shouldn't pay for the drawdown walk.",
        "properties": {
          "beta": {
            "$ref": "#/components/schemas/BetaResult"
          },
          "meta": {
            "$ref": "#/components/schemas/V1Meta"
          }
        },
        "required": [
          "meta",
          "beta"
        ],
        "title": "BetaV1Result",
        "type": "object"
      },
...
      "DrawdownV1Result": {
        "description": "The loss-shaped half of `RiskV1Result`, for consumers that only need\ndrawdown/recovery and shouldn't pay for the beta regression.",
        "properties": {
          "drawdown": {
            "$ref": "#/components/schemas/DrawdownResult"
          },
          "meta": {
            "$ref": "#/components/schemas/V1Meta"
          }
        },
        "required": [
          "meta",
          "drawdown"
        ],
        "title": "DrawdownV1Result",
        "type": "object"
      },
...
      "ExitQualityV1Result": {
        "description": "Sell-side quality facts: regret vs holding and exit alpha vs SPY.",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/V1Meta"
          },
          "result": {
            "$ref": "#/components/schemas/ExitQualityResult"
          }
        },
        "required": [
          "meta",
          "result"
        ],
        "title": "ExitQualityV1Result",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Versioned consumer-read contract for portfolio facts. See docs/api/v1-overview.md for envelope and freshness semantics.",
    "title": "Portfolio Data Service API",
    "version": "1.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analytics/beta": {
      "get": {
        "description": "Beta/alpha/R\u00b2, Sharpe/Sortino, tracking error, volatility only.\n\nThe regression half of `/analytics/risk`, split out so a consumer that\nonly needs volatility does not also pay for the drawdown walk (and vice\nversa \u2014 see `/analytics/drawdown`). `/analytics/risk` still returns both\ntogether for consumers that want one call.",
        "operationId": "analytics_beta_api_v1_analytics_beta_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "benchmark",
            "required": false,
            "schema": {
              "default": "SPY",
              "title": "Benchmark",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "risk_free_annual",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Free Annual"
            }
          },
          {
            "in": "query",
            "name": "exclude_index_etfs",
            "required": false,
            "schema": {
              "default": false,
              "title": "Exclude Index Etfs",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "reserve_amount",
            "required": false,
            "schema": {
              "default": 0.0,
              "minimum": 0,
              "title": "Reserve Amount",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetaV1Result"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analytics Beta",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/analytics/drawdown": {
      "get": {
        "description": "Max drawdown, underwater curve, time-to-recovery, and Calmar only.\n\nThe cheap half of `/analytics/risk`: this walk does not run the beta\nregression, so a consumer needing only loss-shaped risk gets it without\nthe regression's cost.",
        "operationId": "analytics_drawdown_api_v1_analytics_drawdown_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "exclude_index_etfs",
            "required": false,
            "schema": {
              "default": false,
              "title": "Exclude Index Etfs",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "reserve_amount",
            "required": false,
            "schema": {
              "default": 0.0,
              "minimum": 0,
              "title": "Reserve Amount",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrawdownV1Result"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analytics Drawdown",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/analytics/exit-quality": {
      "get": {
        "description": "Sell-side quality facts: regret vs holding and exit alpha vs SPY.",
        "operationId": "analytics_exit_quality_api_v1_analytics_exit_quality_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExitQualityV1Result"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analytics Exit Quality",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/analytics/performance": {
      "get": {
        "description": "Modified-Dietz TWR + benchmark counterfactuals (same calculation as the\nlegacy `/api/portfolio/performance`, now enveloped and versioned).",
        "operationId": "analytics_performance_api_v1_analytics_performance_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "include_backfill",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Backfill",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "reserve_amount",
            "required": false,
            "schema": {
              "default": 0.0,
              "minimum": 0,
              "title": "Reserve Amount",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "exclude_index_etfs",
            "required": false,
            "schema": {
              "default": false,
              "title": "Exclude Index Etfs",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PerformanceV1Result"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analytics Performance",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/analytics/position-performance": {
      "get": {
        "description": "Split-normalized price/trade alpha; unavailable on unresolved share movement.",
        "operationId": "analytics_position_performance_api_v1_analytics_position_performance_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          },
          {
            "in": "query",
            "name": "exclude_broad_index",
            "required": false,
            "schema": {
              "default": false,
              "title": "Exclude Broad Index",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionPerformanceV1Result"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analytics Position Performance",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/analytics/positioning": {
      "get": {
        "operationId": "analytics_positioning_api_v1_analytics_positioning_get",
        "parameters": [
          {
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositioningV1Result"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analytics Positioning",
        "tags": [
          "v1"
        ]
      }
    },
...
  }
}